requirements.txt
asgiref==3.7.2
asttokens==2.4.0
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
decorator==5.1.1
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
executing==1.2.0
Faker==19.3.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipython==8.12.2
jaconv==0.3.4
jedi==0.19.0
kanjize==1.3.0
matplotlib-inline==0.1.6
normalize-japanese-addresses==0.0.6
outcome==1.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pykakasi==2.2.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing-extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
wrapt==1.15.0
wsproto==1.2.0
